Remove old file rewards; Add new files for UpgraderBlock interaction
Removed old file for player's first join rewards; Added new files for the creation and interaction of the UpgraderBlock. Added JSON files for block loot tables, crafting recipes, and new block advancements. Additionally, bump the mod version. This includes adding recipes for smelting, blasting of quartz on different ores and implementing custom smelting logic in the UpgraderBlockEntity, which supports up to 4 simultaneous upgrades. Added texture files for the Dive Upgrade item and updated to include it in the  ItemGroup. This modification also includes the upgrade in Roughly Enough Items and Patchouli's guide book.