Added spear recipe and guidebook entry
A new crafting recipe for a spear item has been added in the Aurus upgrader. It requires a stick, a quartz fragment, an iron ingot, and two copper ingots to craft. Correspondingly, a new entry has been created in the Aurus guidebook providing information about the crafting process and usage of the spear. This information will be displayed in the game to guide the players.